ragify-api/pyproject.toml
[project]
name = "ragify-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt==4.0.1",
    "boto3>=1.40.40",
    "faker>=37.8.0",
    "fastapi>=0.117.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-openai>=0.3.34",
    "openai>=2.1.0",
    "passlib[bcrypt]==1.7.4",
    "psycopg2-binary>=2.9.10",
    "pypdf>=6.1.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.1",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.37.0",
]
